Transformation of Factors to Individual Levels

The function cv.trans.psa takes a covariate data frame and replaces each categorical covariate of n >=3 levels with n new binary covariate columns, one for each level. Transforms covariate dataframe for use with the function cv.bal.psa.

cv.trans.psa(covariates, fcol = NULL)

Arguments

covariates

A dataframe of covariates, presumably some factors.

fcol

An optional vector containing the factor columns in the covariate dataframe. In NULL (default) routine to identfy factors internally.

Value

Returns a dataframe covariates.transformed containing new columns for each level of more than binary factors. The rest of the covariate dataframe stays unchanged.

See also

Author

James E. Helmreich James.Helmreich@Marist.edu

Robert M. Pruzek RMPruzek@yahoo.com

KuangNan Xiong harryxkn@yahoo.com

Examples

#Note reordering of columns, binary factor and numeric column are unchanged.
f2 <- factor(sample(c(0, 1), 20, replace = TRUE))
f4 <- factor(sample(c("a", "b", "c", "d"), 20, replace = TRUE))
cv <- rnorm(20)
X <- data.frame(f2, f4, cv)
cv.trans.psa(X)
#> $covariates.transformed
#>             cv f2_1 f4_a f4_b f4_c f4_d
#> 1  -0.05744799    1    0    1    0    0
#> 2   1.09650605    1    0    0    0    1
#> 3   0.87680026    0    0    0    1    0
#> 4  -0.90489320    0    1    0    0    0
#> 5  -1.43072409    0    0    0    1    0
#> 6   0.82542647    1    0    1    0    0
#> 7   0.51493315    0    0    0    0    1
#> 8   0.85458581    0    0    1    0    0
#> 9  -0.34115402    1    1    0    0    0
#> 10  0.26746070    1    1    0    0    0
#> 11  1.43507539    1    0    0    1    0
#> 12  0.51584666    0    0    0    1    0
#> 13 -1.28008836    0    0    0    0    1
#> 14 -0.47214638    1    0    0    1    0
#> 15  1.41689096    0    0    0    1    0
#> 16 -0.96640941    1    1    0    0    0
#> 17  0.41646346    0    0    0    1    0
#> 18 -1.21898999    0    0    0    1    0
#> 19 -0.36826707    0    0    1    0    0
#> 20  1.06303537    0    0    0    1    0
#> 
#
f2 <- factor(sample(c('c', 'C'), 20, replace = TRUE))
f4 <- factor(sample(c("b", "A", "d", "CC"), 20, replace = TRUE))
cv <- rnorm(20)
X <- data.frame(f2, f4, cv)
cv.trans.psa(X)
#> $covariates.transformed
#>             cv f2_c f4_A f4_CC f4_b f4_d
#> 1   1.08448707    0    0     1    0    0
#> 2   0.03369317    1    1     0    0    0
#> 3   2.17388075    0    1     0    0    0
#> 4   0.82153168    0    0     1    0    0
#> 5   1.55929331    1    0     1    0    0
#> 6   0.04467744    1    0     0    0    1
#> 7  -1.55055429    1    1     0    0    0
#> 8  -0.99769536    0    1     0    0    0
#> 9  -0.03171770    1    0     0    0    1
#> 10  1.42775757    1    0     0    0    1
#> 11  0.81236564    1    0     1    0    0
#> 12  0.61297625    0    1     0    0    0
#> 13  2.76835611    1    0     0    0    1
#> 14  0.25579970    0    1     0    0    0
#> 15  0.19119223    1    0     0    1    0
#> 16  2.62537131    0    0     0    0    1
#> 17  0.30365117    0    0     0    0    1
#> 18 -0.39782499    1    1     0    0    0
#> 19  0.46850138    1    0     0    0    1
#> 20  0.24119770    0    0     0    1    0
#>