Transformation of Factors to Individual Levels

The function cv.trans.psa takes a covariate data frame and replaces each categorical covariate of n >=3 levels with n new binary covariate columns, one for each level. Transforms covariate dataframe for use with the function cv.bal.psa.

cv.trans.psa(covariates, fcol = NULL)

Arguments

covariates

A dataframe of covariates, presumably some factors.

fcol

An optional vector containing the factor columns in the covariate dataframe. In NULL (default) routine to identfy factors internally.

Value

Returns a dataframe covariates.transformed containing new columns for each level of more than binary factors. The rest of the covariate dataframe stays unchanged.

See also

Author

James E. Helmreich James.Helmreich@Marist.edu

Robert M. Pruzek RMPruzek@yahoo.com

KuangNan Xiong harryxkn@yahoo.com

Examples

#Note reordering of columns, binary factor and numeric column are unchanged.
f2 <- factor(sample(c(0, 1), 20, replace = TRUE))
f4 <- factor(sample(c("a", "b", "c", "d"), 20, replace = TRUE))
cv <- rnorm(20)
X <- data.frame(f2, f4, cv)
cv.trans.psa(X)
#> $covariates.transformed
#>               cv f2_1 f4_a f4_b f4_c f4_d
#> 1   2.0738069079    0    0    1    0    0
#> 2   0.6627255919    0    0    0    1    0
#> 3  -0.2577037034    1    0    0    0    1
#> 4   0.0009777376    0    0    0    0    1
#> 5   0.0945484052    0    0    0    1    0
#> 6  -0.1697909027    1    1    0    0    0
#> 7   0.8403645554    0    0    0    1    0
#> 8  -0.1984862211    0    0    0    1    0
#> 9   0.0451910070    0    1    0    0    0
#> 10 -1.4784155442    0    0    0    1    0
#> 11 -0.3812759737    1    0    1    0    0
#> 12 -0.1708654518    1    0    1    0    0
#> 13 -1.3160414117    1    1    0    0    0
#> 14  0.8006089790    0    0    1    0    0
#> 15  0.5335106010    1    0    0    0    1
#> 16 -0.5050495007    1    0    0    1    0
#> 17 -0.7558894476    0    0    0    1    0
#> 18 -0.2112204066    0    0    1    0    0
#> 19 -0.5025069664    0    1    0    0    0
#> 20  0.3618785734    0    0    0    1    0
#> 
#
f2 <- factor(sample(c('c', 'C'), 20, replace = TRUE))
f4 <- factor(sample(c("b", "A", "d", "CC"), 20, replace = TRUE))
cv <- rnorm(20)
X <- data.frame(f2, f4, cv)
cv.trans.psa(X)
#> $covariates.transformed
#>             cv f2_c f4_A f4_CC f4_b f4_d
#> 1   0.56214956    0    0     0    0    1
#> 2   0.41933676    1    0     1    0    0
#> 3   1.04888277    0    0     0    1    0
#> 4  -0.24507283    1    0     1    0    0
#> 5   1.01780346    0    0     0    0    1
#> 6  -0.80178790    1    0     0    0    1
#> 7  -1.33943749    1    1     0    0    0
#> 8   0.83498152    1    0     0    0    1
#> 9   0.17512780    0    1     0    0    0
#> 10 -1.18631930    0    0     0    1    0
#> 11 -0.10917151    1    0     0    1    0
#> 12  2.85842304    1    0     1    0    0
#> 13 -0.37482555    1    0     0    0    1
#> 14 -0.05334079    1    0     0    0    1
#> 15  1.53537773    0    0     1    0    0
#> 16 -0.69873435    0    0     0    0    1
#> 17  1.21540531    1    1     0    0    0
#> 18  1.01982042    1    1     0    0    0
#> 19 -0.17120144    0    0     1    0    0
#> 20  0.03397846    0    0     0    0    1
#>